backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-migrate==4.0.5
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.40
alembic==1.13.1
pandas==2.2.3
openpyxl==3.1.5
argon2-cffi==23.1.0
pyotp==2.9.0
